Rescue exception raised by start_subsegment
AwsExRay.start_subsegment can raise an exception if the ETS table that
it uses has not been initialized. This can happen, for instance, when
running migrations and the AwsExRay.Application.init function is not
called. Rescue the exception so that AwsExRay.Ecto.Logger does not
interfere with running Ecto migrations.